requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
dj-rest-auth==2.2.4
django-allauth==0.47.0
djangorestframework==3.14.0
oauthlib==3.2.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
Pillow==10.1.0
cloudinary==1.37.0
dj3-cloudinary-storage==0.0.6
urllib3
requests==2.31.0
django-filter==23.5